# tests/fixtures/http/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchResultsResponse": {
        "description": "GET /v1/search/results/{sessionId} 响应体.",
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "SearchResultsResponse",
        "type": "object"
      },
      "SearchStatusResponse": {
        "description": "GET /v1/search/status/{sessionId} 响应体.",
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "SearchStatusResponse",
        "type": "object"
      },
      "UnifiedSearchRequest": {
        "description": "POST /v1/search 统一请求体.\n\n智能判断操作类型：\n- 无 sessionId → 新查询（必须有 query）\n- 有 sessionId + query → 追问/继续对话\n- 有 sessionId + 无 query → 恢复历史会话",
        "properties": {
          "location": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
      "description": "位置坐标 {lat, lng}",
      "title": "Location"
    },
    "platforms": {
      "items": {
        "type": "string"
      },
      "type": "array",
      "title": "Platforms",
      "description": "本次允许使用的 MCP 平台"
    },
    "accountRefs": {
      "additionalProperties": {
        "type": "string"
      },
      "type": "object",
      "title": "Accountrefs",
      "description": "按平台选择的账户引用，不包含凭据"
    },
    "expectedSessionVersions": {
      "additionalProperties": {
        "type": "integer"
      },
      "type": "object",
      "title": "Expectedsessionversions",
      "description": "按平台固定的账户会话版本"
    },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "搜索查询（新查询/追问时必填）",
            "title": "Query"
          },
          "sessionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "会话ID（复用现有会话时填写）",
            "title": "Sessionid"
          }
        },
        "title": "UnifiedSearchRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "小红书美食智能推荐 Agent — Production API",
    "title": "XHS Food Agent API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/search/": {
      "post": {
        "description": "Single entry point — branches on ``(sessionId, query)`` presence.",
        "operationId": "unified_search_v1_search__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnifiedSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unified Search",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/search/results/{sessionId}": {
      "get": {
        "operationId": "search_results_v1_search_results__sessionId__get",
        "parameters": [
          {
            "description": "会话ID",
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "description": "会话ID",
              "title": "Sessionid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Results",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/search/status/{sessionId}": {
      "get": {
        "operationId": "search_status_v1_search_status__sessionId__get",
        "parameters": [
          {
            "description": "会话ID",
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "description": "会话ID",
              "title": "Sessionid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Status",
        "tags": [
          "search"
        ]
      }
    },
    "/v1/search/stream/{sessionId}": {
      "get": {
        "description": "Server-sent events with Redis-Stream-backed replay.\n\nThe browser's ``EventSource`` records ``id:`` on every event and sends\nit back as ``Last-Event-ID`` on reconnect. The EventBus replays from\nthat id, so the client needs no special reconnection logic.",
        "operationId": "search_stream_v1_search_stream__sessionId__get",
        "parameters": [
          {
            "description": "会话ID",
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "description": "会话ID",
              "title": "Sessionid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sseVersion",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sseversion"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Stream",
        "tags": [
          "search"
        ]
      }
    },
...
  }
}